Rediseño de la página web aplicando la nueva paleta de colores. Se realizó el cambio de moneda para el peso colombiano, y agrego más productos con tarifas.

+++ server/src/AHuiHou.Application/DTOs/Requests/RedeemPointsRequest.cs
@@ -1,6 +1,6 @@
 namespace AHuiHou.Application.DTOs.Requests;
 
 public record RedeemPointsRequest(
-    int Points
+    decimal Points
 );
