+++ server/src/AHuiHou.Application/DTOs/Responses/PointTransactionResponse.cs
@@ -2,7 +2,7 @@ namespace AHuiHou.Application.DTOs.Responses;
 
 public record PointTransactionResponse(
     Guid Id,
-    int Amount,
+    decimal Amount,
     string TransactionType,
     DateTime CreatedAt
 );
