+++ server/src/AHuiHou.Application/DTOs/Responses/UserResponse.cs
@@ -6,6 +6,6 @@ public record UserResponse(
     string LastName,
     string Email,
     DateTime CreatedAt,
-    int LoyaltyBalance
+    decimal LoyaltyBalance
 );
